+++ CHANGELOG.md
@@ -15,6 +15,8 @@
   fixes spaces in dates to make the commands valid ([rhbz#2163953])
 - Do not display duplicate records in commands `pcs property [config] --all`
   and `pcs property describe` ([rhbz#2217850])
+- Commands `pcs property defaults` and `pcs property describe` print error
+  message in case specified properties do not have metadata. ([rhbz#2222744])
 
 ### Changed
 - When exporting constraints in form of pcs commands, constraints containing
@@ -31,6 +33,7 @@
 [rhbz#2163953]: https://bugzilla.redhat.com/show_bug.cgi?id=2163953
 [rhbz#2217850]: https://bugzilla.redhat.com/show_bug.cgi?id=2217850
 [rhbz#2219407]: https://bugzilla.redhat.com/show_bug.cgi?id=2219407
+[rhbz#2222744]: https://bugzilla.redhat.com/show_bug.cgi?id=2222744
 
 
 ## [0.11.6] - 2023-06-20